requirements.txt
amqp==2.4.2
billiard==3.6.0.0
celery==4.3.0
Django==2.1.8
django-haystack==2.8.1
djangorestframework==3.9.2
drf-haystack==1.8.4
elasticsearch==6.3.1
gunicorn==19.9.0
kombu==4.5.0
Pillow==6.0.0
psycopg2==2.7.7
python-dateutil==2.8.0
python-decouple==3.1
pytz==2018.9
six==1.12.0
sqlparse==0.3.0
urllib3==1.24.1
vine==1.3.0
whitenoise==4.1.2
